+++ mixer/bl_preferences.py
@@ -143,10 +143,12 @@ class MixerPreferences(bpy.types.AddonPreferences):
     send_baked_meshes: bpy.props.BoolProperty(default=True)
 
     display_own_gizmos: bpy.props.BoolProperty(default=False, name="Display Own Gizmos")
-    display_frustums_gizmos: bpy.props.BoolProperty(default=True, name="Display Frustums Gizmos")
-    display_names_gizmos: bpy.props.BoolProperty(default=True, name="Display Name Gizmos")
     display_ids_gizmos: bpy.props.BoolProperty(default=False, name="Display ID Gizmos")
+
+    display_frustums_gizmos: bpy.props.BoolProperty(default=True, name="Display Frustums Gizmos")
+    display_frustums_names_gizmos: bpy.props.BoolProperty(default=True, name="Display Frustums User Names")
     display_selections_gizmos: bpy.props.BoolProperty(default=True, name="Display Selection Gizmos")
+    display_selections_names_gizmos: bpy.props.BoolProperty(default=True, name="Display Selection User Names")
 
     commands_send_interval: bpy.props.FloatProperty(
         name="Command Send Interval",