+++ mixer/bl_properties.py
@@ -145,14 +145,6 @@ class UserItem(bpy.types.PropertyGroup):
     scenes: bpy.props.CollectionProperty(name="Scenes", type=UserSceneItem)
 
 
-# This should probably be handled elsewhere, for now it is here
-# We need a unique index for each user in snap_view_user and snap_time_user
-# dropdown list otherwise the selection "pops" when a user leave room or
-# disconnect
-user_to_unique_index = {}
-next_user_unique_index = 0
-
-
 class SharedFolderItem(bpy.types.PropertyGroup):
     shared_folder: bpy.props.StringProperty(default="", subtype="DIR_PATH", name="Shared Folder")
 
@@ -171,6 +163,7 @@ class MixerProperties(bpy.types.PropertyGroup):
     user_index: bpy.props.IntProperty()  # index in the list of users
 
     display_shared_folders_options: bpy.props.BoolProperty(default=True)
+    display_gizmos_options: bpy.props.BoolProperty(default=True)
     display_advanced_options: bpy.props.BoolProperty(default=False)
     display_developer_options: bpy.props.BoolProperty(default=False)
     display_rooms: bpy.props.BoolProperty(default=True)
@@ -190,80 +183,8 @@ class MixerProperties(bpy.types.PropertyGroup):
     )
     display_users_details: bpy.props.BoolProperty(default=False, name="Display Users Details")
 
-    display_snapping_options: bpy.props.BoolProperty(default=False)
-    snap_view_user_enabled: bpy.props.BoolProperty(default=False)
-
     shared_folders: bpy.props.CollectionProperty(name="Shared Folders", type=SharedFolderItem)
     shared_folder_index: bpy.props.IntProperty()
-
-    def update_user_to_unique_index_dict(self):
-        global user_to_unique_index
-        global next_user_unique_index
-        for user in self.users:
-            if user.ip_port not in user_to_unique_index:
-                user_to_unique_index[user.ip_port] = next_user_unique_index
-                next_user_unique_index += 1
-
-    def get_snap_view_users(self, context):
-        global user_to_unique_index
-
-        self.update_user_to_unique_index_dict()
-
-        # According to documentation:
-        # There is a known bug with using a callback, Python must keep a reference
-        # to the strings returned by the callback or Blender will misbehave or even crash.
-        self.snap_view_users_values = [
-            (user.ip_port, f"{user.name} ({user.ip_port})", "", user_to_unique_index[user.ip_port])
-            for index, user in enumerate(self.users)
-            if user.room == share_data.client.current_room
-        ]
-        return self.snap_view_users_values
-
-    def get_snap_view_area(self, context):
-        # quick patch, see below todo for explanation
-        self.snap_view_users_values = [("id", "", "")]
-        return self.snap_view_users_values
-
-        if self.snap_view_user == "":
-            self.snap_view_areas_values = []
-            return self.snap_view_areas_values
-
-        scene = context.scene
-
-        for user in self.users:
-            if user.ip_port == self.snap_view_user:
-                self.snap_view_areas_values = [
-                    (f"window_{index}", f"Window {index} (Scene {window.scene})", "", index)
-                    for index, window in enumerate(user.windows)
-                    if window.scene == scene.name_full
-                ]
-                return self.snap_view_areas_values
-
-        # According to documentation:
-        # There is a known bug with using a callback, Python must keep a reference
-        # to the strings returned by the callback or Blender will misbehave or even crash.
-        self.snap_view_areas_values = []
-        return self.snap_view_areas_values
-
-    snap_view_user: bpy.props.EnumProperty(
-        items=get_snap_view_users,
-        name="Snap View User",
-    )
-    # todo: this cannot work, it depends on the 3d view panel
-    # todo: so it should be a property of bpy.types.SpaceView3D probably.
-    snap_view_area: bpy.props.EnumProperty(items=get_snap_view_area, name="Snap View 3D Area")
-
-    snap_time_user_enabled: bpy.props.BoolProperty(default=False)
-    snap_time_user: bpy.props.EnumProperty(
-        items=get_snap_view_users,
-        name="Snap Time User",
-    )
-
-    snap_3d_cursor_user_enabled: bpy.props.BoolProperty(default=False)
-    snap_3d_cursor_user: bpy.props.EnumProperty(
-        items=get_snap_view_users,
-        name="Snap 3D Cursor User",
-    )
 
     display_advanced_room_control: bpy.props.BoolProperty(default=False)
     upload_room_name: bpy.props.StringProperty(default=f"{getuser()}_uploaded_room", name="Upload Room Name")